+++ docs/cli/do.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+printf "processing task %s\n" "$1"