+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.0.1-beta.3] – 2022-08-01
+### Added
+* [cli](docs/cli) – now you can run any tasks from the command line
+
 ## [v0.0.1-beta.2] – 2022-07-08
 ### Added
-* Doers
+* [Doers](docs/lib/README.md)
 
 ## [v0.0.1-beta.1] – 2022-07-07
 ### Added
-* TaskManager
+* [TaskManager](docs/lib/README.md) library